requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pdfplumber==0.11.0
python-docx==1.1.0
pytesseract==0.3.10
spacy==3.7.4
google-generativeai==0.4.1
python-dotenv==1.0.1
aiofiles==23.2.1
Pillow==10.3.0
